wip: working on styling

+++ db/migrate/20240731153221_create_nights.rb
@@ -2,7 +2,7 @@ class CreateNights < ActiveRecord::Migration[5.2]
   def change
     create_table :nights do |t|
       t.date :date
-      t.float :cost
+      t.integer :cost
       t.references :trip, foreign_key: true
 
       t.timestamps
